Refactor Vision Dataset Classes and Implement Cross-Entropy Training
- Removed the VisionDataset and PreferenceVisionDataset classes from the vlm/datasets.py file.
- Introduced new cross-entropy training module with VisionDataset and trainer functionalities in the cross-entropy directory.
- Added a new dataset class for preference-based training (ORPO, DPO) in the orpo/datasets.py file.
- Implemented training logic for vision-language models with cross-entropy loss in the cross-entropy/trainer.py file.
- Updated dataset processing methods to handle image and text inputs effectively.
