+++ mlx_vlm/trainer/vlm/cross-entropy/__init__.py
