+++ mlx_vlm/trainer/vlm/cross-entropy/trainer.py
